Modified the "fileRename" method in "FileMethods"
Originally fileRename would have only worked if the extension of a file had a length of 4 (including the '.') because I hard coded the length. I changed the function by using a Path method that retrieves the extension and I use the length of that variable instead. Also added a few more comments for clarity