Added size check for backing up
Before a backup is done, the program now checks the size of the entire backup and checks it against the remaining space on the drive. If the backup size is less than the free space on the drive, it continues with the backup. Otherwise it informs the user that the backup is too large and resets the program to it's opening state